+++ src/reducers/rootReducer.js
@@ -4,9 +4,11 @@ import {combineReducers} from 'redux';
  * Import reducers
  */
 import programState from './programState';
+import dragonflyReducer from './dragonflyReducer';
 
 const rootReducer = combineReducers({
-    programState
+    programState,
+    dragonfly: dragonflyReducer
 });
 
 export default rootReducer;